Still suffering from very high CPU usage in some cases
I'm not sure why this is tripping - before it looked like this was when fullscreen on multi monitors, and DPI scaling was kicking in, but this looks to still be hitting a degenerate case occasionally.

I've put in a temporary fix which could theoretically introduce a small amount of latency.